# Build: Improve `npm run perf` command (fixes #5028)