Fixes #5028.

- A performance test for multi files was added. 
  This test is using `eslint@1.10.3` as the test target.
  This repository has 450 js files in `lib` and `tests/lib` :)
- `.eslintrc` of performance tests came to turn all rules on.

This is new output of this PR:

```
> eslint@2.0.0-beta.2 perf C:\Users\t-nagashima.AD\Documents\GitHub\eslint
> node Makefile.js perf

Loading:
  Load performance Run #1:  193.219655ms
  Load performance Run #2:  193.38688ms
  Load performance Run #3:  191.784973ms
  Load performance Run #4:  191.239836ms
  Load performance Run #5:  203.159501ms

  Load Performance median:  193.219655ms


Single File:
  CPU Speed is 3392 with multiplier 13000000
  Performance Run #1:  3749.119887ms
  Performance Run #2:  3758.879529ms
  Performance Run #3:  3745.368525ms
  Performance Run #4:  3815.086891ms
  Performance Run #5:  3864.7341969999998ms

  Performance budget ok:  3758.879529ms (limit: 3832.5471698113206ms)


Multi Files (450 files):
  CPU Speed is 3392 with multiplier 39000000
  Performance Run #1:  9994.710723ms
  Performance Run #2:  9995.715573ms
  Performance Run #3:  10048.651208ms
  Performance Run #4:  10047.002517ms
  Performance Run #5:  9964.838064ms

  Performance budget ok:  9995.715573ms (limit: 11497.641509433963ms)
```

The loading performance test is the fastest, so I moved it to the first.